This patch fixes IMP-45 and addresses part of IMP-46.
The cause of IMP-45 is a null pointer exception when referencing
HashJoinNode.otherJoinConjuncts. The offending query does not have a "other join
conjuncts".

The fix is to always create an empty list for conjuncts, rather than null. With
an empty list, we don't have do null check everywhere.

IMP-46: I just fix the spacing issue and the duplicate alias. However, the
statement is not a valid SQL. See my comments in the JIRA.

Review: issue 72