This enables running the FE tests in two modes - 'exhaustive' and 'reduced'. Depending on the execution mode, we will use a different set of test vectors so we can help control test execution time. The ideas is that for checkins the tests are run in the 'reduced' input set mode. For nightly builds we will run the exhaustive set of test combinations.
This is controlled with a new flag specified when running the tests:
mvn -DtestExecutionMode=exhaustive test
or
mvn -DtestExecutionMode=reduced test

Note: If the -DtestExecutionMode is not specified it will default to reduced.

As part of this change a bunch of the test files had to be updated to be
parameterized. If they are no parameterized then they will not benefit from the new
coverage that has been added.

This change currently is just for the Query Tests. I would like to extract some
of this logic and generalize it for more test suites with a future checkin.