Enable running Query tests targeting an in-process or out of process (impalad) test env This change enables running of the query tests (and potentially other tests in the future) targeting an in-process or external process test environment. This means that the tests can be run against a remote distributed cluster with ImpalaD deployed - or run locally in-process.
To target a remote environment execute the tests using the following two flags:
mvn test -Dimpalad=<hostname of coordinator> -Dfe_port=21000

If these are not specified that the existing (in-process) test environment is
inferred.

The major parts of this change are:
ImpaladClientExecutor - this is a new client executor class that uses the
beeswax thrift interface to communicate with a target impalad instance.

TestUtilities - This class was updated to add support for running queries
against impalad using the Impalad client executor.

As part of this change I also split the query tests into a few separate files:
JoinQueryTest, InsertQueryTest, HBaseQueryTest, etc... This will make it easier
to pick which subset of tests you want to run. It will also help reduce our max
test log file size in the Jenkins runs.
To enable this I created a new 'BaseQueryTest' class that does much of the work
of choosing which combinations of File format, compression, batch size, etc to
run with.

Current shortcomings:
1) It would be nice for "Executor" and "ImpaladClientExecutor" to share a common
interface. None currently exists and I wasn't sure what a good one would be any
thoughts around this would be appreciated. Because of this I had to resort to
passing an "Exector" of type "Object" for the time being.

2) Beeswax API doesn't currently provide a way to specify things like the number
of execution nodes. For now we just ignore this parameter (it can be set by the
impalad instance).

3) Double and float values are formated with a larger prevision when executed
over the Beeswax interface. This causes results to be different and test failures.
A second checkin will update the in-process output to match that of the beeswax.