IMP- 78
We want to expose issues in an distributed env locally.

We already have 3 data nodes running locally in the MiniDFS. However, the
planner does not distinguish data nodes
on the same host, even though they're running on a different port. So, we're
effectively only running a single node
all the time.

First, we make the change in FE to identify data location as "host/port" instead
of just "host". Then, in
TQueryExecRequest, we list the host/port that serves the data, instead of just
using "host".

The result is that PlannerTest and QueryTest exposes distributed planning issue.
Plans are still correct when the
number of node is 1 or 2. So, to make all the tests passes, I've forced
Planner/Query test to execute with at most 2
nodes.

To see the faulty plan, we simply have to change the number of node back to 0
(all nodes).
o

We've discussed randomizing the SimpleScheduler but I choose not to do it
because we don't need randomization to
expose the distributed planning issue.

I also discovered that exchange node (BE) does not respect the "limit". I fixed
it.

One of the limit test (QueryTest) is completely unstable. It doesn't really test
much. I removed it.