The "combine test/simple scheduler" causes run-query to break.
The reason is that run-query is passing ZERO as the number of nodes to test exec
env, and thus passed to simple scheduler. Simple Scheduler cannot accept ZERO as
the number of nodes.

When FLAGS_num_nodes is set to 1 or 0, it should pass the number of backends
(which is 4 in our MiniCluster config) to the scheduler.