Fix memory issue in hdfs-scan-node.
Row batches produced by the scanner is stored in materialized_row_batches_
and it is consumed by the scan node in GetNext. However, the size of
materialized_row_batches_ is unbounded. It can cause memory spike
if the row batches are not consumed fast enough.

This patch put a size limit on materialized_row_batches_.