IMPALA-27 Restarting an impalad causes queries to hang if it's used as the coordinator
This patches would try to reopen a thrift connection if the thriff
RPC failed. It also changes the "reopen" logic.

When a thrift client reopens, it cannot simply close and open the
undelrying thrift transport: the transport is not properly re-initialized by
"open". Therefore, we are deleting and creating a new thrift client to do a
"reopen".