IMPALA-119 Block locality spew is confusing
In a cluster, we might not have impalad running along side with every
datanode. But that doesn't mean that we have to do remote read.

There are multiple replicas for each block. Therefore we should read
from a replica where it has a collocated impalad first. Only in the case
where none of the replica has a collocated impalad, we'll do remote read.