IMPALA-474: Client in recovery-mode should signal state-store
This fixes a bug where the state-store never detects a client has
failed, because the client is in recovery-mode and ignoring the updates
from the state-store. The state-store therefore never notices the
failure, doesn't expire the subscriber and doesn't allow the subscriber
to re-subscribe (since duplicate 'live' subscriptions are disallowed).

The fix is to make sure that StateStoreSubscriber::UpdateStatus always
returns a representative Status, rather than always returning OK.