Run BE unit tests before all other tests
This change modifies run-all-tests to run the BE unit tests before all other tests.
This is beneficial for a number of reasons - it helps catch basic bugs earlier
(query tests will probably fail if there is a BE unit test failure) and also allows
us to keep the mini-impala-cluster running after a build to help with debugging.
Ideally, we could also run the FE unit tests before the query tests but there is
currently a dependency on the TPC-H temp tables generated by the query tests so this
cannot be done.

Change-Id: Id43dbac456236258cd9986e990779d27f5d41075
Reviewed-on: http://gerrit.ent.cloudera.com:8080/269
Tested-by: jenkins
Tested-by: Ishaan Joshi <ishaan@cloudera.com>
Reviewed-by: Ishaan Joshi <ishaan@cloudera.com>