IMPALA-564: Close HS2 sessions on socket close
This patch adds a SessionHandlerIf to the HiveServer2 Thrift
server. When a socket connection event happens, Impala will terminate
the associated HS2 sessions.

Since HS2 allows for multiple sessions to be multiplexed onto the same
socket connection, a list of session IDs associated with each connection
ID is added to ImpalaServer so that all sessions can be correctly
terminated. This patch also fixes the HS2 implementation so that all
sessions on the same socket get different session IDs.

This patch also adds two metrics to track the number of sessions
currently active for both HS2 and Beeswax APIs.

Change-Id: Ic6276c86b0ab842ac9f434afccd14ca49937bee8
Reviewed-on: http://gerrit.ent.cloudera.com:8080/343
Tested-by: jenkins
Reviewed-by: Alan Choi <alan@cloudera.com>