IMPALA-592: Improve performance of ALTER TABLE ADD PARTITION
ALTER TABLE ADD PARTITION had performance problems as it scaled to a
large number of partitions. This was because Impala was always

The reason Impala is hit this is because after executing the
ALTER DDL, Impala "refreshes" the table metadata. As part of the
refresh() Impala tries to reuse any metadata it can that already is
cached. In this case the lastDdlTime has changed so all partitions are
reloaded from the metastore using the listPartitions() RPC. This call
does not scale well as the number of partitions grows, for a table with
2000 partitions ADD PARTITION can take over 20 seconds.

This change significantly improves the performance of ALTER TABLE
ADD PARTITION by slightly changing how incremental refreshes work. We
now check the list of partition names in the metastore and load only the
delta of what is new (and remove any partitions that have been dropped)
by checking a new "isDirty()" flag on HdfsPartition. Additionally,
the lastDdlTime is now updated on the local (cached) copy after each
ALTER operation so we can detect against external modification to the
table.

With these changes we are able to add partitions at a pretty much
constant time (~1s / partition), even for tables that have a large
number of partitions.

Change-Id: Idc48618d4061ea3c56d9b6dae2c431a7ac49d5d9
Reviewed-on: http://gerrit.ent.cloudera.com:8080/495
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>