IMPALA-588: Reset col stats to unknown instead of failing when col stats data is incompatible
Instead of failing when Hive ColumnStatsData isn't compatible with the
column type, Impala will now reset that Column's stats to "unknown".
This allows the table metadata to be loaded, for what could be a common
scenario - someone computes stats on a column and then then changes the
column type using an ALTER TABLE command.

Change-Id: I7f681b8e6b6d35268a84794014912122d1fefab6
Reviewed-on: http://gerrit.ent.cloudera.com:8080/515
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins