IMPALA-568: Retry Register() statestore subscriber RPCs
Before introducing a more sophisticated client-cache, we used to have
the ability to open Thrift clients with retry and timeout
parameters. This patch adds that back so that all clients drawn from a
single cache will have the same retry behaviour. This is used only in
the statestore subscriber for now, which needs to retry since the
statestore might come up after an Impala demon has started.

Change-Id: Ifb2f6456d7dcbecd40c79293e43b72b585d2cf37
Reviewed-on: http://gerrit.ent.cloudera.com:8080/527
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>