Add CHAR(N) type in BE.
It's going to be pretty suboptimal to implement UDAs without having CHAR(N)
support so I implemented the bare minimum to support it. We need to change
all uses of PrimitiveType in the future.

This is a bit hard to test now since we don't expose this in the language except
for UDAs currently. I can combine this with the UDA patch but that patch is
pretty big.

Change-Id: I799dd2c905b41194e92cc01728727546294b0a02
Reviewed-on: http://gerrit.ent.cloudera.com:8080/562
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>