IMPALA-470 (part 2/2): Expire idle queries
This patch adds individual query expiration. Queries are expired if they
are 'idle' for too long. A query is idle if it has neither been actively
processed (by e.g. FetchInternal()) nor received a client
command (e.g. Fetch()) in the last FLAGS_idle_query_timeout seconds.

The general idea is to keep a priority queue of query expiration times
which is walked periodically by a thread to expire those queries whose
expiration time is in the past. In order to reduce locking, queries do
not directly update their expiration times in that queue. Instead, the
queue keeps an estimate of the expiration time which is guaranteed to be
a lower bound, and then wakes up when that time has passed to see
whether it is still accurate. Therefore there is minimal lock
interaction between the expiration logic, and the core query processing
flow.

For a longer explanation, see impala-server.h and in particular
ExpirationQueue.

Change-Id: I954d33d0484eefaacb205cb7ee352b950900ef14
Reviewed-on: http://gerrit.ent.cloudera.com:8080/631
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: jenkins