Gather catalog updates in a separate thread from the state store heartbeat thread
Instead of gathering catalog updates in the same thread as the state store heartbeat
thread, this change introduces a separate thread that gathers updates from
the JniCatalog and stores them to a shared data structure. This ensures the actual time
spent processing a heartbeat by the catalog server stays small. The second thread
is signaled to start at end of every heartbeat and the heartbeat thread will simply
do a no-op while the second thread is executing.

Change-Id: Ie412898cb1dd4e7a049d2535eee4f2dece6b72ee
Reviewed-on: http://gerrit.ent.cloudera.com:8080/674
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>
Tested-by: Lenni Kuff <lskuff@cloudera.com>