IMPALA-613: Ensure no table ID conflicts between tables in a query
Normally the catalog service assigns all table IDs, but in the case of CTAS
Impalad create a local "temp" table. This "temp" table was getting assigned
an ID that conflicted with a real table. The fix is to just ensure the ID is
unique within all tables in the query, so assign the target table in the CTAS
statement a table ID of -1. Once the table is actually created in the catalog
server it will be assigned a proper table ID.

Change-Id: I21551354fca40d09e2155aee94b52c9aca97028c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/729
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: jenkins